refactor: structure of main page and inner components
