+++ src/App.jsx
@@ -1,13 +1,12 @@
 import Header from "./components/Header";
-import ActorPage from "./pages/ActorPage";
-import { person } from "./mock-data/mockPersonDetails";
+import MainPage from "./pages/MainPage";
 import "./App.scss";
 
 const App = () => {
   return (
     <div className="App">
       <Header />
-      <ActorPage {...person} />
+      <MainPage></MainPage>
     </div>
   );
 };
