+++ src/pages/MainPage/MainPageContainer.jsx
@@ -1,5 +1,5 @@
-import MovieCardList from "../../../components/MovieCardList";
-import DataFromTMDB from "../../../mock-data/mockMovies";
+import MovieCardList from "../../components/MovieCardList";
+import DataFromTMDB from "../../mock-data/mockMovies";
 const MainPageContainer = () => {
   // todo: add mockData from MovieDB
   return <MovieCardList {...DataFromTMDB} />;
